[code-clean] handle common issues reported by pylint
This patch adds a pylint profile (.pylintrc) with the aim to have a good
compromise in context of Python v2 & v3 and the status of the current
implementation.

To avoid having to switch off all checks in the profile, many of the issues
reported by pylint were handled by using inline comments.

Simple errors in the code were corrected, but only if the effect of the change
was locally limited and manageable.

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>