[code-clean] The datetime class is a subclass of date.
Since the class `datetime.datetime` is a subclass of `datetime.date`, the
`isinstance` check is not suitable if a distinction has to be made between the
two classes.

The `type` check is more suitable in these cases, as no further inheritance can
be assumed for these two object classes and the code is also easier to read in
the context of this distinction.

The situation is completely different with method:

- icalendar.Trigger.transformFromNative()

in this method (in the Trgger class) a distinction is made between a date and a
time duration.  Since `datetime.date` is the base class, it is suitable for an
`isinstance` check of a date object (which can be an instance of the
`datetime.datetime` class as well as `datetime.date`).

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>