[hatch] replace Makefile by Hatch, a project management tool
Hatch [1] is a project management tool for the python eco system.  With Hatch
all project definitions are stored in pyproject.toml.  The Makefile and the
requirements.txt files have been removed.

In the github-CI we have a quality gate that runs tests on python:

- 2.7, 3.7, 3.8, 3.9, 3.10, 3.11 and 3.12

Developers can test modifications by::

   $ hatch run dev:pre-commit

To run tests in a matrix of Python versions locally, developers can use:

   $ hatch run quality-gate

About installation and usage read chapter "Development" in the README.md file.

[1] https://hatch.pypa.io

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>